move ayon utils widgets to utils
