--- client/ayon_core/tools/ayon_utils/widgets/__init__.py
@@ -1,51 +1,0 @@
-from .projects_widget import (
-    # ProjectsWidget,
-    ProjectsCombobox,
-    ProjectsQtModel,
-    ProjectSortFilterProxy,
-    PROJECT_NAME_ROLE,
-    PROJECT_IS_CURRENT_ROLE,
-    PROJECT_IS_ACTIVE_ROLE,
-    PROJECT_IS_LIBRARY_ROLE,
-)
-
-from .folders_widget import (
-    FoldersWidget,
-    FoldersQtModel,
-    FOLDERS_MODEL_SENDER_NAME,
-    SimpleFoldersWidget,
-)
-
-from .tasks_widget import (
-    TasksWidget,
-    TasksQtModel,
-    TASKS_MODEL_SENDER_NAME,
-)
-from .utils import (
-    get_qt_icon,
-    RefreshThread,
-)
-
-
-__all__ = (
-    # "ProjectsWidget",
-    "ProjectsCombobox",
-    "ProjectsQtModel",
-    "ProjectSortFilterProxy",
-    "PROJECT_NAME_ROLE",
-    "PROJECT_IS_CURRENT_ROLE",
-    "PROJECT_IS_ACTIVE_ROLE",
-    "PROJECT_IS_LIBRARY_ROLE",
-
-    "FoldersWidget",
-    "FoldersQtModel",
-    "FOLDERS_MODEL_SENDER_NAME",
-    "SimpleFoldersWidget",
-
-    "TasksWidget",
-    "TasksQtModel",
-    "TASKS_MODEL_SENDER_NAME",
-
-    "get_qt_icon",
-    "RefreshThread",
-)
