+++ client/ayon_core/tools/utils/__init__.py
@@ -56,6 +56,28 @@ from .dialogs import (
     SimplePopup,
     PopupUpdateKeys,
 )
+from .projects_widget import (
+    ProjectsCombobox,
+    ProjectsQtModel,
+    ProjectSortFilterProxy,
+    PROJECT_NAME_ROLE,
+    PROJECT_IS_CURRENT_ROLE,
+    PROJECT_IS_ACTIVE_ROLE,
+    PROJECT_IS_LIBRARY_ROLE,
+)
+
+from .folders_widget import (
+    FoldersWidget,
+    FoldersQtModel,
+    FOLDERS_MODEL_SENDER_NAME,
+    SimpleFoldersWidget,
+)
+
+from .tasks_widget import (
+    TasksWidget,
+    TasksQtModel,
+    TASKS_MODEL_SENDER_NAME,
+)
 
 
 __all__ = (
@@ -115,4 +137,21 @@ __all__ = (
     "ScrollMessageBox",
     "SimplePopup",
     "PopupUpdateKeys",
+
+    "ProjectsCombobox",
+    "ProjectsQtModel",
+    "ProjectSortFilterProxy",
+    "PROJECT_NAME_ROLE",
+    "PROJECT_IS_CURRENT_ROLE",
+    "PROJECT_IS_ACTIVE_ROLE",
+    "PROJECT_IS_LIBRARY_ROLE",
+
+    "FoldersWidget",
+    "FoldersQtModel",
+    "FOLDERS_MODEL_SENDER_NAME",
+    "SimpleFoldersWidget",
+
+    "TasksWidget",
+    "TasksQtModel",
+    "TASKS_MODEL_SENDER_NAME",
 )
