+++ client/ayon_core/tools/utils/projects_widget.py
@@ -1,7 +1,8 @@
 from qtpy import QtWidgets, QtCore, QtGui
 
 from ayon_core.tools.ayon_utils.models import PROJECTS_MODEL_SENDER
-from .utils import RefreshThread, get_qt_icon
+
+from .lib import RefreshThread, get_qt_icon
 
 PROJECT_NAME_ROLE = QtCore.Qt.UserRole + 1
 PROJECT_IS_ACTIVE_ROLE = QtCore.Qt.UserRole + 2
